testsuite, c++, contracts: Update tests for modified parser.
This updates a few of the erorr messages - it also fixes up cases where
the user gave conflicting modifiers (to ignore them) so that the baseline
constification takes place.

All the tests are updated to use the p2900 syntax, but a copy of the sub-
set of tests that were using attribute (or mixed) syntax has been kept
in cpp26-attr.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>