testsuite: arm: Use effective-target for its.c test [PR94531]
The test case gcc.target/arm/its.c was created together with restriction
of IT blocks for Cortex-M7. As the test case fails on all tunes that
does not match Cortex-M7, explicitly test it for Cortex-M7. To have some
additional faith that GCC does the correct thing, I also added another
variant of the test for Cortex-M3 that should allow longer IT blocks.

gcc/testsuite/ChangeLog:

	PR testsuite/94531
	* gcc.target/arm/its.c: Removed.
	* gcc.target/arm/its-1.c: Copy of gcc.target/arm/its.c. Use
	effective-target arm_cpu_cortex_m7.
	* gcc.target/arm/its-2.c: Copy of gcc.target/arm/its.c. Use
	effective-target arm_cpu_cortex_m3.

Signed-off-by: Torbjörn SVENSSON <torbjorn.svensson@foss.st.com>