ada: Ensure minimum stack size for preallocated task stacks
On targets with preallocated task stacks the minimum stack size is
defined as a constant in System.Parameters. When adding preallocated
tasks to the expanded code the compiler does not have direct access to
that value. Instead generate the expression
Max (Task_Size, Minimum_Task_Size) in the expanded tree and let it be
resolved later in the compilation process.

gcc/ada/ChangeLog:

	* exp_ch9.adb (Expand_N_Task_Type_Declaration): Take
	Minimum_Stack_Size into account when preallocating task stacks.
	* rtsfind.ads (RE_Id, RE_Unit_Table): Add RE_Minimum_Stack_Size.