cfgexpand: Skip doing conflicts if there is only 1 variable
This is a small speed up. If there is only one know stack variable, there
is no reason figure out the scope conflicts as there are none. So don't
go through all the live range calculations just to see there are none.

Bootstrapped and tested on x86_64-linux-gnu with no regressions.

gcc/ChangeLog:

	* cfgexpand.cc (add_scope_conflicts): Return right away
	if there are only one stack variable.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>