ada: Minor adjustments to error message for RM B.1(24)
The RM B.1(24) sub-clause says that imported entities cannot be initialized
and it is checked in three contexts, aspect Import, pragma Import and pragma
Import_Object, with slightly different error messages.  Moreover, for the
aspect, the error is given twice because that of the pragma is also given.

In addition, if the initialization expression is an aggregate that is not
static, the error is given only for the aspect and not for the two pragmas.

This change aligns the error messages on that of pragma Import and plugs the
aforementioned loophole for the two pragmas.

gcc/ada/ChangeLog:

	* sem_ch13.adb (Analyze_Aspect_Export_Import): Add explicit mention
	of the declaration in the error message for the Import.
	* sem_prag.adb (Process_Extended_Import_Export_Object_Pragma): Also
	test Has_Init_Expression on the declaration node for Import_Object
	and use the same wording as that of Import.
	(Process_Import_Or_Interface): Also test Has_Init_Expression on the
	declaration node for Import.